refactor: meta merger
