+++ src/commons/index.ts
@@ -1,3 +1,4 @@
 export { ArrayUtilities } from './array-utilities.js';
 export { HashUtilities } from './hash-utilities.js';
+export { MetaMerger } from './meta-merger.js';
 export { StringUtilities } from './string-utilities.js';
